+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 ## Unreleased
 ### Fixed
 - Incorrect condition when checking instance on ObjectHelper::create
+- Trait conflict with ElementView::runInternal
 
 ## 1.0.0-rc.15 - 2018-03-20
 ### Added